pyproject.toml
[project]
name = "template-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "erwan serandour", email = "e.serandour@hispeed.ch" }
]
requires-python = ">=3.12"
dependencies = [
    "dspy>=3.0.4",
    "fastapi[standard]>=0.127.0",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
]

[project.scripts]
template-tools = "template_tools:main"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"
